requirements.txt
certifi==2020.6.20
chardet==3.0.4
Django==2.2
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
idna==2.10
Markdown==3.2.2
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
